# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GlutenFree documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/products/store": {
            "post": {
                "tags": [
                    "Admin-Products"
                ],
                "summary": "Create product",
                "description": "!!! use form-data for api !!!. Example of request: https://drive.google.com/file/d/1lNZPNG8k4ieZeR4Zohj1Ci61s92vZafP/view?usp=sharing",
                "operationId": "store-product",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/products/update/{slug}": {
            "post": {
                "tags": [
                    "Admin-Products"
                ],
                "summary": "Update product by slug",
                "description": "!!! use form-data for api !!!. Images 'is_main' field is not working. Example of request: https://drive.google.com/file/d/1lNZPNG8k4ieZeR4Zohj1Ci61s92vZafP/view?usp=sharing",
                "operationId": "update-product",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Product slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "moloko-moloczyj-mir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/products/delete/{slug}": {
            "post": {
                "tags": [
                    "Admin-Products"
                ],
                "summary": "Delete product",
                "description": "Delete product with images",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Product slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "moloko-moloczyj-mir"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/admin/product-subcategories/store": {
            "post": {
                "tags": [
                    "Admin-Products-Subcategories"
                ],
                "summary": "Create products subcategory",
                "description": "",
                "operationId": "store-product-subcategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductSubcategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/product-subcategories/update/{slug}": {
            "post": {
                "tags": [
                    "Admin-Products-Subcategories"
                ],
                "summary": "Update products subcategory",
                "description": "",
                "operationId": "update-product-subcategory",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Products-subcategory slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "jogurt"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductSubcategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/admin/product-subcategories/delete/{slug}": {
            "post": {
                "tags": [
                    "Admin-Products-Subcategories"
                ],
                "summary": "Delete products subcategory",
                "description": "",
                "operationId": "delete-product-subcategory",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Products-subcategory slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "jogurt"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get all products",
                "description": "Return all products with category, subcategory and images, paginated by 12",
                "operationId": "products",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Product category slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "molochnyje-produkty"
                    },
                    {
                        "name": "subcategory",
                        "in": "query",
                        "description": "Product subcategory slug",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "moloko"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": "List of products. Pagination by 12"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "title": "StorePostRequest",
                "description": "Class StorePostRequest",
                "required": [
                    "subcategory_id",
                    "map_id",
                    "lang",
                    "name",
                    "images"
                ],
                "properties": {
                    "subcategory_id": {
                        "title": "subcategory_id",
                        "description": "Products subcategory ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "map_id": {
                        "title": "map_id",
                        "description": "Products map ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "lang": {
                        "title": "lang",
                        "description": "language of post",
                        "type": "string",
                        "example": "pl"
                    },
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Chleb"
                    },
                    "description": {
                        "title": "description",
                        "description": "product description",
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 1,
                        "example": "Вкусный хлеб"
                    },
                    "maker": {
                        "title": "maker",
                        "description": "product maker",
                        "type": "string",
                        "maxLength": 255,
                        "minLength": 1,
                        "example": "Гродненский хлебозавод"
                    },
                    "images": {
                        "title": "images",
                        "description": "array of images",
                        "type": "string",
                        "example": "['image' => image, 'is_main' => 1]"
                    }
                },
                "type": "object"
            },
...
            "StoreProductSubcategoryRequest": {
                "title": "StoreProductSubcategoryRequest",
                "description": "Class StoreProductSubcategoryRequest",
                "required": [
                    "category_id",
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "Id of products category",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of products subcategory",
                        "type": "string",
                        "maxLength": 180,
                        "minLength": 2,
                        "example": "Jogurt"
                    }
                },
                "type": "object"
            },
            "UpdateProductSubcategoryRequest": {
                "title": "UpdateProductSubcategoryRequest",
                "description": "Class UpdateProductSubcategoryRequest",
                "required": [
                    "name"
                ],
                "properties": {
                    "category_id": {
                        "title": "category_id",
                        "description": "Id of products category",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "Name of products subcategory",
                        "type": "string",
                        "maxLength": 180,
                        "minLength": 2,
                        "example": "Jogurt"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer \"token\")",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        []
    ]
}